Attribute hits to throwers

+++ dys-simulation/src/combatant_statline.rs
@@ -6,6 +6,6 @@ pub struct CombatantStatline {
     pub combatant_id: CombatantId,
     pub points_scored: u8,
     pub balls_thrown: u16,
-    pub throws_hit: u8,
+    pub throws_hit: u16,
     pub combatants_shoved: u16,
 }
